study 도메인의 dto를 하나로 합쳐라
study 도메인의 dto를 하나로 합칩니다. study와 관련된 도메인들도 합쳐야
합니다
