--- app/src/main/java/com/example/bookclub/dto/StudyUpdateDto.java
@@ -1,58 +1,0 @@
-package com.example.bookclub.dto;
-
-import com.example.bookclub.domain.Day;
-import com.example.bookclub.domain.StudyState;
-import com.example.bookclub.domain.Zone;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateDeserializer;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class StudyUpdateDto {
-    private String name;
-
-    private String description;
-
-    private String contact;
-
-    private int size;
-
-    @JsonDeserialize(using = LocalDateDeserializer.class)
-    private LocalDate startDate;
-
-    @JsonDeserialize(using = LocalDateDeserializer.class)
-    private LocalDate endDate;
-
-    private String startTime;
-
-    private String endTime;
-
-    private Day day;
-
-    private StudyState studyState;
-
-    private Zone zone;
-
-    @Builder
-    public StudyUpdateDto(String name, String description, String contact,
-                          int size, LocalDate startDate, LocalDate endDate, String startTime,
-                          String endTime, Day day, StudyState studyState, Zone zone) {
-        this.name = name;
-        this.description = description;
-        this.contact = contact;
-        this.size = size;
-        this.startDate = startDate;
-        this.endDate = endDate;
-        this.startTime = startTime;
-        this.endTime = endTime;
-        this.day = day;
-        this.studyState = studyState;
-        this.zone = zone;
-    }
-}
